requirements.txt
boto3==1.34.113          # AWS SDK for S3 interaction
pandas==2.2.2            # Core library for data manipulation and cleaning
pyarrow==16.1.0          # For efficient Parquet file creation
gspread==6.1.1           # Python API for Google Sheets
oauth2client==4.1.3      # Google Sheets authentication (Service Account)
psycopg2-binary==2.9.9   # Postgres database connector
python-dotenv==1.0.1     # To securely load environment variables from .env
SQLAlchemy==2.0.30       # Python SQL toolkit (optional, but good practice)